Reworking the message file.
There is a new tool, genmsg, that is capable of generating multiple
formats of message files for different platforms via a common
source, specifically src/tecomsg.src.

The syntax is documented in genmsg.c source file and the example can
be found in the source mentioned above.

At the moment it is possible to generate OpenVMS HELP and MESSAGE
utility files.  There is basic support for the header file that holds
the message constants.  However, this will need further work as will
the teco_putmsg routine which still needs adjustment.